add rest of views and create viewset
